#32: Add capability to write QuarkusIntegrationTests
While at it, reformat and unify code.
