+++ deployment/src/test/java/io/quarkiverse/cucumber/test/BasicTest.java
@@ -3,5 +3,4 @@ package io.quarkiverse.cucumber.test;
 import io.quarkiverse.cucumber.CucumberQuarkusTest;
 
 public class BasicTest extends CucumberQuarkusTest {
-
 }
