+++ deployment/src/test/java/io/quarkiverse/cucumber/test/Steps.java
@@ -11,23 +11,24 @@ import io.cucumber.java.en.Then;
 import io.restassured.response.ValidatableResponse;
 
 public class Steps {
+    private final String target;
 
     @Inject
-    @ConfigProperty(defaultValue = "/")
-    String target;
+    public Steps(@ConfigProperty(name = "some.endpoint.path", defaultValue = "/") String target) {
+        this.target = target;
+    }
 
     private ValidatableResponse result;
 
     @Given("I call the endpoint")
-    public void i_call_endpoint() throws Exception {
+    public void i_call_endpoint() {
         result = given()
                 .when().get(target)
                 .then();
     }
 
     @Then("the response is ok")
-    public void response_is_ok() throws Exception {
+    public void response_is_ok() {
         result.statusCode(200);
     }
-
 }
