+++ integration-tests/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-testPath=/cucumber
+greeting="Hello, cucumber!"
