+++ integration-tests/src/test/java/io/quarkiverse/cucumber/it/CucumberResourceIT.java
@@ -1,0 +1,9 @@
+package io.quarkiverse.cucumber.it;
+
+import io.quarkiverse.cucumber.CucumberQuarkusIntegrationTest;
+
+public class CucumberResourceIT extends CucumberQuarkusIntegrationTest {
+    public static void main(String... args) {
+        runMain(CucumberResourceIT.class, args);
+    }
+}
