+++ integration-tests/src/test/java/io/quarkiverse/cucumber/it/CucumberResourceTest.java
@@ -3,7 +3,7 @@ package io.quarkiverse.cucumber.it;
 import io.quarkiverse.cucumber.CucumberQuarkusTest;
 
 public class CucumberResourceTest extends CucumberQuarkusTest {
-    public static void main(String[] args) {
+    public static void main(String... args) {
         runMain(CucumberResourceTest.class, args);
     }
 }
